+++ actionpack/CHANGELOG.md
@@ -1,3 +1,10 @@
+*   Introduce `config.action_controller.dom_testing_assertions` and `config.action_dispatch.dom_testing_assertions`
+
+    Adds support for `:rails_dom_testing` to support `Rails::Dom::Testing::Assertions` and `:none`.
+    Defaults to with `:rails_dom_testing`.
+
+    *Sean Doyle*
+
 *   Include default headers in `ActionController::Live` responses.
 
     Previously, responses from `ActionController::Live` controllers, including